Revert "merged legacy hooks tests"
This reverts commit 924a83dec6e153dda88ab3b75066e86bfd4e9435.
